Actions work  ActionsResolver init method @PostConstruct

+++ src/main/java/ru/ivan/sqlcmd/controller/actions/DropTableAction.java
@@ -23,18 +23,22 @@ public class DropTableAction extends AbstractAction {
 
     @Override
     public void get(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        DatabaseManager manager = getManager(req, resp);
-        String tableName = req.getParameter("table");
-        try {
-            manager.dropTable(tableName);
-            req.setAttribute("message", String.format("Table '%s' dropped successfully!",
-                    tableName));
-            req.setAttribute("link", "tables");
-            req.setAttribute("title", "Back to tables list");
-            goToJsp(req, resp, "message.jsp");
-        } catch (Exception e) {
-            req.setAttribute("message", String.format("Table '%s' cannot be dropped!", tableName));
-            goToJsp(req, resp, "error.jsp");
+        DatabaseManager manager = (DatabaseManager) req.getSession().getAttribute("manager");
+        if (manager != null) {
+            String tableName = req.getParameter("table");
+            try {
+                manager.dropTable(tableName);
+                req.setAttribute("message", String.format("Table '%s' dropped successfully!",
+                        tableName));
+                req.setAttribute("link", "tables");
+                req.setAttribute("title", "Back to tables list");
+                goToJsp(req, resp, "message.jsp");
+            } catch (Exception e) {
+                req.setAttribute("message", String.format("Table '%s' cannot be dropped!", tableName));
+                goToJsp(req, resp, "error.jsp");
+            }
+        } else {
+            resp.sendRedirect(resp.encodeRedirectURL("connect"));
         }
     }
 }
